requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
pdfplumber==0.10.0
python-docx==0.8.11
Pillow==10.1.0
pytesseract==0.3.10
google-generativeai==0.3.2
openai==1.12.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0

jinja2==3.1.2
gradio==4.8.0
aiofiles==23.2.1